V1.0.4 Redesign of projects page + basic scroll animations
Redesigned the projects page to replace slideshow with clean display of projects. Important projects get more space on the page.
Added a back to top button at the bottom right which fades in at a certain scroll point - on education, experience and projects pages.
Animated hash scrolling for back to top button and for the secondary navigation controls on the experience, education and projects pages.